Cache sighashes throughout a scripts execution
In CHECKMULTISIG, signatures are repeatedly tried
against public keys until the correct order is found.
The transaction signature hashing algorithm is run
for each signature, often leading to the same hash
being calculated each time.

A method CalculateSignatureHash is exposed on
TransactionSignatureChecker, with a caching variant
for CachingTransactionSignatureChecker which uses
the new CSignatureHashCache.